requirements.txt
edge-tts==7.0.2
pydub==0.25.1
ffmpeg-python
opencv-python>=4.9.0.80
easyocr==1.7.1
numpy>=1.26.4
torch==2.2.2
scipy>=1.12.0
Pillow
piexif
pillow-heif
requests 
flask